services/configuration-management/requirements.txt
# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database
psycopg2-binary==2.9.9
asyncpg==0.29.0
PyJWT==2.8.0
sqlalchemy==2.0.23
alembic==1.13.1

# Redis
redis==5.0.1
hiredis==2.2.3

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Configuration & Environment
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0

# Monitoring & Logging
prometheus-client==0.19.0
structlog==23.2.0

# HTTP Client
httpx==0.25.2

# Validation & Serialization
email-validator==2.1.0

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1